split methods.py
