+++ ephios/core/signup/methods.py
@@ -1,42 +1,26 @@
-import dataclasses
 import functools
 import logging
 from argparse import Namespace
 from collections import OrderedDict
-from typing import List, Optional
+from typing import List
 
-from crispy_forms.bootstrap import FormActions
-from crispy_forms.helper import FormHelper
-from crispy_forms.layout import HTML, Field, Layout
-from django import forms
-from django.contrib import messages
-from django.core.exceptions import PermissionDenied, ValidationError
-from django.db import transaction
+from django.core.exceptions import ValidationError
 from django.db.models import Q
-from django.shortcuts import redirect
 from django.template.loader import get_template
 from django.urls import reverse
 from django.utils import timezone
-from django.utils.formats import date_format
-from django.utils.functional import SimpleLazyObject, cached_property, classproperty
-from django.utils.timezone import localtime
+from django.utils.functional import cached_property
 from django.utils.translation import gettext_lazy as _
-from django.views.generic import FormView
 
 from ephios.core.dynamic_preferences_registry import GeneralRequiredQualificationPreference
 from ephios.core.models import AbstractParticipation, Shift
 from ephios.core.services.notifications.types import (
-    ResponsibleConfirmedParticipationCustomizedNotification,
     ResponsibleConfirmedParticipationDeclinedNotification,
 )
-from ephios.core.signals import (
-    check_participant_signup,
-    participant_from_request,
-    register_signup_methods,
-)
+from ephios.core.signals import check_participant_signup, register_signup_methods
 from ephios.core.signup.participants import AbstractParticipant
+from ephios.core.signup.stats import SignupStats
 from ephios.extra.utils import format_anything
-from ephios.extra.widgets import CustomSplitDateTimeWidget
 
 logger = logging.getLogger(__name__)
 
@@ -89,252 +73,6 @@ class ImproperlyConfiguredError(ParticipationError):
     Error to return if signup cannot be performed for
     technical or configuration reasons.
     """
-
-
-def prevent_getting_participant_from_request_user(request):
-    """
-    To prevent plugin authors from accessing the participant using request.user, the SignupView
-    uses this method to block access to `as_participant` of the request.user attribute value.
-    """
-    original_user = request.user
-
-    class ProtectedUser(SimpleLazyObject):
-        def as_participant(self):
-            raise AttributeError(
-                "Access of request.user.as_participant in SignupViews is not allowed."
-            )
-
-    request.user = ProtectedUser(lambda: original_user)
-
-
-def get_nonlocal_participant_from_request(request):
-    for _, participant in participant_from_request.send(sender=None, request=request):
-        if participant is not None:
-            return participant
-    raise PermissionDenied
-
-
-class BaseParticipationForm(forms.ModelForm):
-    individual_start_time = forms.SplitDateTimeField(
-        label=_("Individual start time"), widget=CustomSplitDateTimeWidget, required=False
-    )
-    individual_end_time = forms.SplitDateTimeField(
-        label=_("Individual end time"),
-        widget=CustomSplitDateTimeWidget,
-        required=False,
-    )
-
-    def clean_individual_start_time(self):
-        if self.cleaned_data["individual_start_time"] == self.shift.start_time:
-            return None
-        return self.cleaned_data["individual_start_time"]
-
-    def clean_individual_end_time(self):
-        if self.cleaned_data["individual_end_time"] == self.shift.end_time:
-            return None
-        return self.cleaned_data["individual_end_time"]
-
-    def clean(self):
-        cleaned_data = super().clean()
-        if not self.errors:
-            start = cleaned_data["individual_start_time"] or self.shift.start_time
-            end = cleaned_data["individual_end_time"] or self.shift.end_time
-            if end < start:
-                self.add_error("individual_end_time", _("End time must not be before start time."))
-            return cleaned_data
-
-    class Meta:
-        model = AbstractParticipation
-        fields = ["individual_start_time", "individual_end_time", "comment"]
-
-    def __init__(self, *args, **kwargs):
-        instance = kwargs["instance"]
-        kwargs["initial"] = {
-            **kwargs.get("initial", {}),
-            "individual_start_time": instance.individual_start_time or self.shift.start_time,
-            "individual_end_time": instance.individual_end_time or self.shift.end_time,
-        }
-        super().__init__(*args, **kwargs)
-
-    def get_customization_notification_info(self):
-        """
-        Return a list of human readable messages for changed participation attributes responsibles should be informed about.
-        This should not include the participation state, but customization aspects such as individual times, detailed disposition information, etc.
-        """
-        assert self.is_valid()
-        info = []
-        for time in ["start_time", "end_time"]:
-            if (field_name := f"individual_{time}") in self.changed_data:
-                info.append(
-                    _("{label} was changed from {initial} to {current}.").format(
-                        label=self.fields[field_name].label,
-                        initial=date_format(
-                            localtime(self.initial[field_name].astimezone()),
-                            format="SHORT_DATETIME_FORMAT",
-                        ),
-                        current=date_format(
-                            localtime(self.cleaned_data[field_name] or getattr(self.shift, time)),
-                            format="TIME_FORMAT",
-                        ),
-                    )
-                )
-
-        return info
-
-
-class BaseSignupForm(BaseParticipationForm):
-    def _get_field_layout(self):
-        return Layout(*(Field(name) for name in self.fields))
-
-    def _get_buttons(self):
-        if (
-            p := self.participant.participation_for(self.method.shift)
-        ) is not None and p.is_in_positive_state():
-            buttons = [
-                HTML(
-                    f'<button class="btn btn-success mt-1 ms-1 float-end" type="submit" name="signup_choice" value="customize">{_("Save")}</button>'
-                )
-            ]
-        else:
-            buttons = [
-                HTML(
-                    f'<button class="btn btn-success mt-1 ms-1 float-end" type="submit" name="signup_choice" value="sign_up">{self.method.registration_button_text}</button>'
-                )
-            ]
-        buttons.append(
-            HTML(
-                f'<a class="btn btn-secondary mt-1" href="{self.participant.reverse_event_detail(self.method.shift.event)}">{_("Cancel")}</a>'
-            )
-        )
-        if self.method.can_decline(self.participant):
-            buttons.append(
-                HTML(
-                    f'<button class="btn btn-secondary mt-1 ms-1 float-end" type="submit" name="signup_choice" value="decline">{_("Decline")}</button>'
-                )
-            )
-        return buttons
-
-    def __init__(self, *args, **kwargs):
-        self.method = kwargs.pop("method")
-        self.targets_positive_state = kwargs.pop("targets_positive_state")
-        self.shift: Shift = self.method.shift
-        self.participant: AbstractParticipant = kwargs.pop("participant")
-        super().__init__(*args, **kwargs)
-        self.helper = FormHelper()
-        self.helper.layout = Layout(
-            self._get_field_layout(),
-            FormActions(*self._get_buttons()),
-        )
-
-        if not self.shift.signup_method.configuration.user_can_customize_signup_times:
-            self.fields["individual_start_time"].disabled = True
-            self.fields["individual_end_time"].disabled = True
-
-    def clean(self):
-        cleaned_data = super().clean()
-        self._validate_conflicting_participations(self, self.targets_positive_state)
-        return cleaned_data
-
-    def _validate_conflicting_participations(self, form, targets_positive_state):
-        if not targets_positive_state:
-            return
-        if conflicts := get_conflicting_participations(
-            participant=form.instance.participant,
-            start_time=form.cleaned_data.get("individual_start_time") or self.shift.start_time,
-            end_time=form.cleaned_data.get("individual_end_time") or self.shift.end_time,
-            shift=self.shift,
-            total=False,
-        ):
-            form.add_error("individual_start_time", "")
-            form.add_error(
-                "individual_end_time",
-                _("You are already confirmed for other shifts at this time: {shifts}.").format(
-                    shifts=", ".join(str(shift) for shift in conflicts)
-                ),
-            )
-
-
-class BaseSignupView(FormView):
-    """
-    This View reacts to the signup or decline buttons being pressed using a POST request.
-    Beware that request.user might be anonymous. You should only act on participants acquired
-    using `get_participant`.
-    """
-
-    shift: Shift = ...
-    method: "BaseSignupMethod" = ...
-    form_class = BaseSignupForm
-    template_name = "core/signup.html"
-
-    def get_form_kwargs(self):
-        kwargs = super().get_form_kwargs()
-        kwargs.update(
-            {
-                "method": self.method,
-                "participant": self.participant,
-                "instance": self.participation,
-                "targets_positive_state": self.request.POST.get("signup_choice") == "sign_up",
-            }
-        )
-        return kwargs
-
-    @cached_property
-    def participation(self):
-        return self.method.get_participation_for(self.participant)
-
-    def setup(self, request, *args, **kwargs):
-        super().setup(request, *args, **kwargs)
-        self.participant: AbstractParticipant = kwargs["participant"]
-        prevent_getting_participant_from_request_user(request)
-
-    def form_valid(self, form):
-        if (choice := self.request.POST.get("signup_choice")) is not None:
-            if choice == "sign_up" and self.method.can_sign_up(self.participant):
-                return self.signup_pressed(form)
-            if choice == "customize" and self.method.can_customize_signup(self.participant):
-                return self.customize_pressed(form)
-            if choice == "decline":
-                return self.decline_pressed(form)
-            messages.error(self.request, _("This action is not allowed."))
-            return redirect(self.participant.reverse_event_detail(self.shift.event))
-        form.add_error(None, _("Form submission is missing the mode of signup."))
-        return self.form_invalid(form)
-
-    def customize_pressed(self, form):
-        form.save()
-        if claims := form.get_customization_notification_info():
-            ResponsibleConfirmedParticipationCustomizedNotification.send(form.instance, claims)
-        messages.success(self.request, _("Your participation was saved."))
-        return redirect(self.participant.reverse_event_detail(self.shift.event))
-
-    def signup_pressed(self, form):
-        try:
-            with transaction.atomic():
-                participation = form.save()
-                self.method.perform_signup(self.participant, participation, **form.cleaned_data)
-        except ParticipationError as errors:
-            for error in errors:
-                messages.error(self.request, self.method.signup_error_message.format(error=error))
-        else:
-            messages.success(
-                self.request,
-                self.method.signup_success_message.format(shift=self.shift),
-            )
-        return redirect(self.participant.reverse_event_detail(self.shift.event))
-
-    def decline_pressed(self, form):
-        try:
-            with transaction.atomic():
-                participation = form.save()
-                self.method.perform_decline(self.participant, participation, **form.cleaned_data)
-        except ParticipationError as errors:
-            for error in errors:
-                messages.error(self.request, self.method.decline_error_message.format(error=error))
-        else:
-            messages.info(
-                self.request, self.method.decline_success_message.format(shift=self.shift)
-            )
-        return redirect(self.participant.reverse_event_detail(self.shift.event))
 
 
 def check_event_is_active(method, participant):
@@ -450,33 +188,6 @@ def check_conflicting_participations(method, participant):
         )
 
 
-class BaseSignupMethodConfigurationForm(forms.Form):
-    template_name = "core/signup_configuration_form.html"
-    minimum_age = forms.IntegerField(
-        required=False, min_value=1, max_value=999, initial=None, label=_("Minimum age")
-    )
-    signup_until = forms.SplitDateTimeField(
-        required=False,
-        widget=CustomSplitDateTimeWidget,
-        initial=None,
-        label=_("Signup until"),
-    )
-    user_can_decline_confirmed = forms.BooleanField(
-        label=_("Confirmed users can decline by themselves"),
-        required=False,
-        help_text=_("only if the signup timeframe has not ended"),
-    )
-    user_can_customize_signup_times = forms.BooleanField(
-        label=_("Users can provide individual start and end times"),
-        required=False,
-        initial=True,
-    )
-
-    def __init__(self, *args, **kwargs):
-        self.event = kwargs.pop("event")
-        super().__init__(*args, **kwargs)
-
-
 class BaseSignupMethod:
     # pylint: disable=too-many-public-methods
 
@@ -498,10 +209,16 @@ class BaseSignupMethod:
 
         return BaseDispositionParticipationForm
 
-    configuration_form_class = BaseSignupMethodConfigurationForm
+    @property
+    def configuration_form_class(self):
+        from ephios.core.signup.forms import BaseSignupMethodConfigurationForm
+
+        return BaseSignupMethodConfigurationForm
 
     @property
     def signup_view_class(self):
+        from ephios.core.signup.views import BaseSignupView
+
         return BaseSignupView
 
     @property
@@ -767,46 +484,3 @@ class BaseSignupMethod:
             kwargs.setdefault("event", self.event)
         form = self.configuration_form_class(*args, **kwargs)
         return form
-
-
-@dataclasses.dataclass(frozen=True)
-class SignupStats:
-    requested_count: int
-    confirmed_count: int
-    missing: int
-    free: Optional[int]  # None means infinite free
-    min_count: Optional[int]  # None means no min specified
-    max_count: Optional[int]  # None means infinite max
-
-    @classproperty
-    def ZERO(cls):  # pylint: disable=no-self-argument
-        return SignupStats(
-            requested_count=0,
-            confirmed_count=0,
-            missing=0,
-            free=0,
-            min_count=None,
-            max_count=0,
-        )
-
-    def __add__(self, other: "SignupStats"):
-        free = self.free + other.free if self.free is not None and other.free is not None else None
-        missing = self.missing + other.missing
-        min_count = (
-            (self.min_count or 0) + (other.min_count or 0)
-            if self.min_count is not None or other.min_count is not None
-            else None
-        )
-        max_count = (
-            self.max_count + other.max_count
-            if self.max_count is not None and other.max_count is not None
-            else None
-        )
-        return SignupStats(
-            requested_count=self.requested_count + other.requested_count,
-            confirmed_count=self.confirmed_count + other.confirmed_count,
-            missing=missing,
-            free=free,
-            min_count=min_count,
-            max_count=max_count,
-        )
